Add test for pipe item applying charmed effect
Tests are added to check that the magical pipe item correctly
applies the Charmed.java attack strategy to enemies. It checks that
enemies become charmed when a player equips the pipe, and return to
normal when the player unequips it. It also checks that new enemies
immediately become charmed when they target a player already holding
the pipe.

change file: MagicalPipeTest.java