Add charmed attack strategy
The canAttackNow() method in the stub Charmed.java attack strategy is
overrided to always return false. This means any creature with this
strategy will never be able to attack its target. The charmed strategy
is also added to AttackStrategyFactory.java so it be added to an
entity.

change file: Charmed.java, AttackStrategyFactory.java

Related issue: #11