Fix test for charmed when equipped
The setup for the charmed when equipped test has been updated to
correctly simulate the player equipping an item. When equipping an
item with PlayerTestHelper, the onEquipped() method for the item was
never called. To fix this, onEquipped() & onUnequipped() are manually
called within the test. Also the creatures are made to attack their
targets before the pipes are equipped to ensure they are in the attack
sources list.

change file: MagicalPipeTest.java

Related issue: #11