Fix error when planting seeds in garden
The player cannot plant seeds in garden initially because garden is not
a fertile ground. I edit the Shovel.java file to use shovel to configure
the ground where the shovel is as a fertile ground. We can put shovel in
garden to make it a fertile ground and the player can then plant seeds
on it.

Issue: #15